// src/transaction/enum/transaction.enum.ts
export enum CourseStatus {
  ADD_SUBJECT = 'ADD_SUBJECT',
  WITHDRAWAL = 'WITHDRAWAL',
  CHANGE_COURSE = 'CHANGE_COURSE',
  CHANGE_SECTION = 'CHANGE_SECTION',
}

export enum TypeDocument {
  RO01 = 'RO-01 คำร้องทั่วไป',
  RO16 = 'RO-16 คำร้องขอลาป่วย ลากิจ',
  RO26 = 'RO-26 ใบลงทะเบียนเพิ่ม-ลด-ถอน-เปลี่ยนกลุ่ม-เปลี่ยนรายวิชา',
}